Changed interpolation rate value to q8_24
This is so we can easily add an offset to the interpolation rate.

Also, it is more often required to have more fractional precision than
integral precision. With the current format a wavetable can be played at just
over 128 times its original rate, but with 2^24 bits of fractional
precision.

The tests have also been updated to reflect the change.